T4generator - excluding helper function to additional file
